# [WIP] perf: remove TrieUpdates::removed_nodes