--- a/bin/reth/src/commands/debug_cmd/in_memory_merkle.rs
+++ b/bin/reth/src/commands/debug_cmd/in_memory_merkle.rs
@@ -197,9 +197,9 @@ impl<C: ChainSpecParser<ChainSpec = ChainSpec>> Command<C> {
         // Compare updates
         let mut in_mem_mismatched = Vec::new();
         let mut incremental_mismatched = Vec::new();
-        let mut in_mem_updates_iter = in_memory_updates.account_nodes_ref().iter().peekable();
+        let mut in_mem_updates_iter = in_memory_updates.changed_nodes_ref().iter().peekable();
         let mut incremental_updates_iter =
-            incremental_trie_updates.account_nodes_ref().iter().peekable();
+            incremental_trie_updates.changed_nodes_ref().iter().peekable();
 
         while in_mem_updates_iter.peek().is_some() || incremental_updates_iter.peek().is_some() {
             match (in_mem_updates_iter.next(), incremental_updates_iter.next()) {