--- a/crates/trie/sparse/src/state.rs
+++ b/crates/trie/sparse/src/state.rs
@@ -16,7 +16,11 @@ use reth_trie_common::{
     MultiProof, MultiProofTargets, Nibbles, RlpNode, TrieAccount, TrieNode, EMPTY_ROOT_HASH,
     TRIE_ACCOUNT_RLP_MAX_SIZE,
 };
-use std::{collections::VecDeque, fmt, iter::Peekable};
+use std::{
+    collections::{HashMap, VecDeque},
+    fmt,
+    iter::Peekable,
+};
 
 /// Sparse state trie representing lazy-loaded Ethereum state trie.
 pub struct SparseStateTrie<F: BlindedProviderFactory = DefaultBlindedProviderFactory> {
@@ -454,8 +458,10 @@ impl<F: BlindedProviderFactory> SparseStateTrie<F> {
         self.state.as_revealed_mut().map(|state| {
             let updates = state.take_updates();
             TrieUpdates {
-                account_nodes: updates.updated_nodes,
-                removed_nodes: updates.removed_nodes,
+                changed_nodes: HashMap::from_iter(Iterator::chain(
+                    updates.removed_nodes.into_iter().map(|k| (k, None)),
+                    updates.updated_nodes.into_iter().map(|(k, v)| (k, Some(v))),
+                )),
                 storage_tries: self
                     .storages
                     .iter_mut()
@@ -742,7 +748,7 @@ mod tests {
         pretty_assertions::assert_eq!(
             sparse_updates,
             TrieUpdates {
-                account_nodes: HashMap::default(),
+                changed_nodes: HashMap::default(),
                 storage_tries: HashMap::from_iter([(
                     b256!("1100000000000000000000000000000000000000000000000000000000000000"),
                     StorageTrieUpdates {
@@ -751,7 +757,6 @@ mod tests {
                         removed_nodes: HashSet::default()
                     }
                 )]),
-                removed_nodes: HashSet::default()
             }
         );
     }