- [x] Benchmarks

The basic idea is to combine one `HashMap<Nibbles, BranchNodeCompact>` and one `HashSet<Nibbles>` into one `HashMap<Nibbles, Option<BranchNodeCompact>>` so that the hot operations can be optimized.

## Notable changes

```diff
pub struct TrieUpdates {
-    pub account_nodes: HashMap<Nibbles, BranchNodeCompact>,
-    pub removed_nodes: HashSet<Nibbles>,
+    pub changed_nodes: HashMap<Nibbles, Option<BranchNodeCompact>>,
     pub storage_tries: B256HashMap<StorageTrieUpdates>,
 }

pub struct StorageTrieUpdates {
     pub is_deleted: bool,
-    pub storage_nodes: HashMap<Nibbles, BranchNodeCompact>,
-    pub removed_nodes: HashSet<Nibbles>,
+    pub changed_nodes: HashMap<Nibbles, Option<BranchNodeCompact>>,
 }
 ```

## Benchmarks


![Total run time for state root calculation (1)](https://github.com/user-attachments/assets/5233c62e-4ab8-47c5-87b3-32c0629af88a)



<details>
<summary>Benchmarks</summary>

------
### Summary


  | Before | After | Ratio
-- | -- | -- | --
ERC20 | 585111076 | 510748357 | 0.8729083724
Raw Transfer | 323457659 | 293110568 | 0.9061790928
Uniswap | 1254878698 | 1139740260 | 0.9082473564

(unit: microseconds)



### Before

- erc20-from16-1914-super-low-dependency.zip (30600/150000) <br> 30,601.00 tps, 1,056,914,565.83 gps, no chain lag, 585111076/609686114 μs
- raw-transfer-from05-1723-super-low-dependency.zip (54600/280000) <br> 54,594.73 tps, 1,146,512,812.78 gps, no chain lag, 323457659/332195794 μs
- uniswap-from14-1848-super-low-dependency.zip (6330/25320) <br> 6,331.00 tps, 882,451,082.10 gps, no chain lag, 1254878698/1270738569 μs

### After

- erc20-from16-1914-super-low-dependency.zip (30600/150000) <br> 30,601.00 tps, 1,056,918,521.38 gps, no chain lag, 510748357/534427714 μs
- raw-transfer-from05-1723-super-low-dependency.zip (54600/280000) <br> 54,601.00 tps, 1,146,644,512.40 gps, no chain lag, 293110568/302033002 μs
- uniswap-from14-1848-super-low-dependency.zip (6330/25320) <br> 6,331.00 tps, 882,444,901.42 gps, no chain lag, 1139740260/1157216500 μs

### Others

```
[183548e8:3c34fef1]

"erc20-from16-1914-super-low-dependency.zip"
{'tps': 33001.0, 'gps': 1139819457.5775578, 'is_chain_lagged': False, 'chain_lag_distance': 0}
{'tps': 33101.0, 'gps': 1143258673.245033, 'is_chain_lagged': False, 'chain_lag_distance': 1}
{'tps': 33301.0, 'gps': 1150171044.6133332, 'is_chain_lagged': False, 'chain_lag_distance': 1}
{'tps': 33701.0, 'gps': 1163995862.5292134, 'is_chain_lagged': True, 'chain_lag_distance': 2}
{'tps': 34501.0, 'gps': 1191624258.379747, 'is_chain_lagged': False, 'chain_lag_distance': 1}
{'tps': 36101.0, 'gps': 1246888006.3610108, 'is_chain_lagged': True, 'chain_lag_distance': 14}

"raw-transfer-from05-1723-super-low-dependency.zip"
{'tps': 58001.0, 'gps': 1218044525.1627908, 'is_chain_lagged': True, 'chain_lag_distance': 2}
{'tps': 58201.0, 'gps': 1222244509.7359223, 'is_chain_lagged': True, 'chain_lag_distance': 3}
{'tps': 58601.0, 'gps': 1230644515.4794521, 'is_chain_lagged': True, 'chain_lag_distance': 6}
{'tps': 59401.0, 'gps': 1247444525.3333333, 'is_chain_lagged': True, 'chain_lag_distance': 7}
{'tps': 61001.0, 'gps': 1281044511.0367346, 'is_chain_lagged': True, 'chain_lag_distance': 13}

"uniswap-from14-1848-super-low-dependency.zip"
{'tps': 6578.742763157895, 'gps': 916967104.8754386, 'is_chain_lagged': False, 'chain_lag_distance': 0}
{'tps': 6616.001543890604, 'gps': 922155973.8782532, 'is_chain_lagged': True, 'chain_lag_distance': 5}
{'tps': 6721.94622451266, 'gps': 936937043.6361192, 'is_chain_lagged': True, 'chain_lag_distance': 23}

[785bc168:b19d00f9]

"erc20-from16-1914-super-low-dependency.zip"
{"tps": 30501.0, "gps": 1053468679.4730417, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 31501.0, "gps": 1088008368.302521, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 33001.0, "gps": 1139817415.4389439, "is_chain_lagged": true, "chain_lag_distance": 2}
{"tps": 36001.0, "gps": 1243429312.1416566, "is_chain_lagged": true, "chain_lag_distance": 11}

"raw-transfer-from05-1723-super-low-dependency.zip"
{"tps": 53601.0, "gps": 1125644516.6236558, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 56001.0, "gps": 1176044516.6067417, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 58001.0, "gps": 1218044507.8491297, "is_chain_lagged": true, "chain_lag_distance": 2}

[785bc168:4a2d9bda]

"erc20-from16-1914-super-low-dependency.zip"
{"tps": 33001.0, "gps": 1139820803.359736, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 34001.0, "gps": 1174356683.510204, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 35751.0, "gps": 1234791949.373063, "is_chain_lagged": true, "chain_lag_distance": 2}
{"tps": 36001.0, "gps": 1243435230.1584635, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 36001.0, "gps": 1243426739.6902761, "is_chain_lagged": true, "chain_lag_distance": 2}
{"tps": 36251.0, "gps": 1252064356.1620314, "is_chain_lagged": true, "chain_lag_distance": 2}
{"tps": 36751.0, "gps": 1269330520.0735295, "is_chain_lagged": true, "chain_lag_distance": 4}
{"tps": 37751.0, "gps": 1303869778.450189, "is_chain_lagged": true, "chain_lag_distance": 11}
{"tps": 39751.0, "gps": 1372950063.740053, "is_chain_lagged": true, "chain_lag_distance": 32}
{"tps": 40001.0, "gps": 1381594207.2496662, "is_chain_lagged": true, "chain_lag_distance": 28}

"raw-transfer-from05-1723-super-low-dependency.zip"
{"tps": 53601.0, "gps": 1125644514.3434706, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 56001.0, "gps": 1176044515.8202248, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 58001.0, "gps": 1218044505.5512574, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 58001.0, "gps": 1218044508.7674417, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 58251.0, "gps": 1223294505.2451363, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 58716.427450980394, "gps": 1233068493.8823528, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 59751.0, "gps": 1254794510.7250996, "is_chain_lagged": true, "chain_lag_distance": 4}
{"tps": 61751.0, "gps": 1296794518.7505155, "is_chain_lagged": true, "chain_lag_distance": 7}

"uniswap-from14-1848-super-low-dependency.zip"
{"tps": 6581.0, "gps": 917307342.6959193, "is_chain_lagged": true, "chain_lag_distance": 6}
{"tps": 6601.0, "gps": 920076335.2767876, "is_chain_lagged": true, "chain_lag_distance": 20}
{"tps": 6641.0, "gps": 925652474.0219123, "is_chain_lagged": true, "chain_lag_distance": 11}
{"tps": 6721.0, "gps": 936810861.6591978, "is_chain_lagged": true, "chain_lag_distance": 43}
{"tps": 6881.0, "gps": 959113335.8423951, "is_chain_lagged": true, "chain_lag_distance": 118}
{"tps": 7201.0, "gps": 1003732445.9757562, "is_chain_lagged": true, "chain_lag_distance": 284}

[785bc168:95be8fb6]

"erc20-from16-1914-super-low-dependency.zip"
{"tps": 33001.0, "gps": 1139815818.1188118, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 34001.0, "gps": 1174353299.537415, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 35001.0, "gps": 1208894706.0373394, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 36001.0, "gps": 1243432276.429772, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 36001.0, "gps": 1243429342.3793519, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 36001.0, "gps": 1243427352.6410565, "is_chain_lagged": true, "chain_lag_distance": 3}
{"tps": 36401.0, "gps": 1257249328.1407766, "is_chain_lagged": true, "chain_lag_distance": 2}
{"tps": 36801.0, "gps": 1271065507.1141105, "is_chain_lagged": true, "chain_lag_distance": 5}
{"tps": 37001.0, "gps": 1277963946.7814815, "is_chain_lagged": true, "chain_lag_distance": 6}
{"tps": 37201.0, "gps": 1284878472.1637716, "is_chain_lagged": true, "chain_lag_distance": 6}

"raw-transfer-from05-1723-super-low-dependency.zip"
{"tps": 58401.0, "gps": 1226444512.994152, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 58721.0, "gps": 1233164517.552941, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 59041.0, "gps": 1239884519.929134, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 59361.0, "gps": 1246604509.2380953, "is_chain_lagged": false, "chain_lag_distance": 1}
{"tps": 60001.0, "gps": 1260044515.490982, "is_chain_lagged": false, "chain_lag_distance": 0}
{"tps": 61001.0, "gps": 1281044511.355102, "is_chain_lagged": true, "chain_lag_distance": 3}
{"tps": 62001.0, "gps": 1302044522.1490684, "is_chain_lagged": true, "chain_lag_distance": 8}
{"tps": 63001.0, "gps": 1323044515.487395, "is_chain_lagged": true, "chain_lag_distance": 12}
```

------
</details>